+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## next
 
+- Store the initial image data in the right place for avoiding [side-effect](https://github.com/fengyuanchen/jquery-cropper/issues/19).
 - Fix the issue when the value of the `minContainerWidth/Height` options is `0` (#683).
 
 ## 1.5.7 (May 23, 2020)